55: extract `<RequestPasswordReset>` into the newly-created `frontend/src/features/auth/` directory
